Split daily-data into type, query, and result files for better organization.
