+++ __tests__/helpers/daily-data-providers/combined-therapy-minutes.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from '@jest/globals';
 import { createEmptyCombinedDataCollectionSettings, createMockResult, sampleEndDate, sampleResult, sampleStartDate, setupCombinedDataCollectionSettings, setupCombinedFirstValueResult, setupDailyDataProvider } from '../../fixtures/daily-data-providers';
 import combinedTherapyMinutes from '../../../src/helpers/daily-data-providers/combined-therapy-minutes';
-import * as dailyDataFunctions from '../../../src/helpers/daily-data-providers/daily-data';
+import * as dailyDataResultFunctions from '../../../src/helpers/daily-data-providers/daily-data/daily-data-result';
 import { appleHealthTherapyMinutesDataProvider, googleFitTherapyMinutesDataProvider } from '../../../src/helpers/daily-data-providers';
 
 jest.mock('../../../src/helpers/daily-data-providers/apple-health-therapy-minutes', () => ({
@@ -18,7 +18,7 @@ describe('Daily Data Provider - Combined Therapy Minutes', () => {
 
     const appleHealthTherapyMinutesDataProviderMock = appleHealthTherapyMinutesDataProvider as jest.Mock;
     const googleFitTherapyMinutesDataProviderMock = googleFitTherapyMinutesDataProvider as jest.Mock;
-    const combinedFirstValueResultMock = jest.spyOn(dailyDataFunctions, 'combineResultsUsingFirstValue');
+    const combinedFirstValueResultMock = jest.spyOn(dailyDataResultFunctions, 'combineResultsUsingFirstValue');
 
     beforeEach(() => {
         jest.resetAllMocks();
